fix: download moved to file card
